@Antonov-Alexander 2023-01-14
TestGetDomainStat_Time_And_Memory
*Old*
```
    stats_optimization_test.go:45: time used: 573.152001ms / 300ms
    stats_optimization_test.go:46: memory used: 307Mb / 30Mb
```

*New*
```
    stats_optimization_test.go:46: time used: 146.677154ms / 300ms
    stats_optimization_test.go:47: memory used: 4Mb / 30Mb
```